+++ src/scorer/feed/diversity_feed_scorer.ts
@@ -3,7 +3,7 @@
  * prevent prolific tooters from clogging up the feed.
  */
 import FeedScorer from "../feed_scorer";
-import Toot from '../../api/objects/toot';
+import Toot, { sortByCreatedAt } from '../../api/objects/toot';
 import { incrementCount, shuffle } from "../../helpers/collection_helpers";
 import { StringNumberDict, WeightName } from "../../types";
 
@@ -13,11 +13,15 @@ export default class DiversityFeedScorer extends FeedScorer {
         super(WeightName.DIVERSITY);
     }
 
+    // Count toots by account (but negative instead of positive count)
+    // *NOTE: The penalty for frequent tooters decreases by 1 each time a toot is scored*
+    //        As a result this.scoreData must be reset anew each time the feed is updated
     feedExtractor(feed: Toot[]): StringNumberDict {
-        // Shuffle the feed to avoid biasing the scoring based on the order of the feed
-        // Count toots by account (but negative instead of positive count)
-        // TODO: maybe reverse chronological order would be better?
-        return shuffle<Toot>(feed).reduce(
+        const sortedToots = sortByCreatedAt(feed).reverse() as Toot[];
+        console.info(`DiversityFeedScorer: ${sortedToots.length} toots in sorted feed:`, sortedToots.slice(0, 10));
+
+        // Collate the overall score for each account
+        const accountScores = sortedToots.reduce(
             (tootCounts, toot) => {
                 incrementCount(tootCounts, toot.account.webfingerURI(), -1);
                 if (toot.reblog?.account) incrementCount(tootCounts, toot.reblog.account.webfingerURI(), -1);
@@ -25,23 +29,27 @@ export default class DiversityFeedScorer extends FeedScorer {
             },
             {} as StringNumberDict
         );
+
+        // Create a dict with a score for each toot, keyed by uri (mutates userScores)
+        return sortedToots.reduce(
+            (scores, toot) => {
+                incrementCount(accountScores, toot.account.webfingerURI());
+                scores[toot.uri] = accountScores[toot.account.webfingerURI()] || 0;
+
+                if (toot.reblog) {
+                    incrementCount(accountScores, toot.reblog.account.webfingerURI());
+                    scores[toot.uri] += (accountScores[toot.reblog.account.webfingerURI()] || 0);
+                }
+
+                return scores;
+            },
+            {} as StringNumberDict
+        );
     }
 
-    // *NOTE: The penalty for frequent tooters decreases by 1 each time a toot is scored*
-    //        As a result this.features must be reset anew each time the feed is scored
     async _score(toot: Toot) {
-        incrementCount(this.scoreData, toot.account.webfingerURI());
-        if (toot.reblog) incrementCount(this.scoreData, toot.reblog.account.webfingerURI());
-        const acct = toot.realAccount().webfingerURI();
-
-        // TODO: this was a hack to avoid wildly overscoring diversity values because of a bug that should be fixed now
-        if (this.scoreData[acct] > 0) {
-            let msg = `DiversityFeedScorer for ${toot.account.webfingerURI()} scored over 0`;
-            msg += ` (got ${this.scoreData[toot.account.webfingerURI()]})`;
-            console.warn(`${msg}, diversity features:\n${JSON.stringify(this.scoreData, null, 4)}`);
-            return 0;
-        } else {
-            return this.scoreData[acct];
-        }
+        const score = this.scoreData[toot.uri] || 0;
+        if (score > 0) console.warn(`Got positive diversity score of ${score} for toot:`, toot);
+        return score;
     }
 };