+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # NEXT RELEASE
+* Use reverse chronological order sort in DiversityFeedScorer; stop mutating DiversityFeedScorer.scoreData during scoring
+* Score toots with Promises instead of serially, add `Config.scoringBatchSize` param
 
 ### v0.20.4
 * Add `Config.staleDataSeconds` param, consolidate `Storage.isStaleData()` method